Major refactor for better architectured components

+++ app/src/main/java/com/pesimatik/appbox/model/ActivityUsagesRepository.kt
@@ -6,19 +6,19 @@ import java.io.ObjectInputStream
 import java.io.ObjectOutputStream
 
 object ActivityUsagesRepository {
-    fun getUsages(context: Context): Map<String, Int> {
+    fun getUsages(context: Context): Map<String, Float> {
         val f = File(context.filesDir, "usages")
         return if (!f.exists()) {
             f.createNewFile()
             emptyMap()
         } else {
             f.inputStream().use {
-                ObjectInputStream(it).readObject() as Map<String, Int>
+                ObjectInputStream(it).readObject() as Map<String, Float>
             }
         }
     }
 
-    fun storeUsages(context: Context, usages: Map<String, Int>) {
+    fun storeUsages(context: Context, usages: Map<String, Float>) {
         File(context.filesDir, "usages").outputStream().use {
             ObjectOutputStream(it).writeObject(usages)
         }
